Renamed test/utils folder
